+++ TumorGrowthToolkit/FK_DTI/__init__.py
@@ -1,1 +1,1 @@
-from .FK import Solver
+from .FK_DTI import Solver
